requirements.txt
cx_Freeze==7.1.1
pillow==10.3.0
pystray==0.19.5
pywin32==306
requests==2.32.3
winshell==0.6